feature/MMT-1821 and feature/MMT-1407 updated from feedback from pull request.
